change theme

+++ assets/script/theme.js
@@ -6,11 +6,11 @@ let themeModelLoc = "<div id=\"theme-ID\" class=\"theme-tab text-center\">" +
 	"<input type=\"button\" class=\"theme-del\" value=\"删除\"><br>" +
 	"<label class=\"theme-name\">THEMENAME</label>" +
 	"</div>";
-let themeGlbBg = ["white", "white", "#FFF7D3", "#F5E8C8", "#F5F1DA", "#E0D5F2", "#FEEEED", "#6C757D", "#7F9BA8"];
-let themeGlbSub = ["red", "skyblue", "#00768F", "#AE1027", "#808C6C", "#0E97B6", "#EF5B9C", "#6C757D", "#D0D3DC"];
-let themeGlbCd = ["red", "skyblue", "#FAC457", "#A12F2F", "#F4B46A", "#0E97B6", "#EF5B9C", "#6C757D", "#C8B493"];
-let themeGlbTxt = ["black", "black", "#00768F", "#752423", "#9A9C94", "black", "#F7ACBC", "#6C757D", "#383C4A"];
-let themeGlbName = ["经典", "经典2", "游乐场(投稿)", "JasonYing都觉得好(投稿)", "春天来啦~(投稿)", "奶思(投稿)", "樱花粉(投稿)", "你看个屁(投稿)", "海盐夏日(投稿)"];
+let themeGlbBg = ["white", "white", "black", "#FFF7D3", "#F5E8C8", "#F5F1DA", "#E0D5F2", "#FEEEED", "#6C757D", "#7F9BA8"];
+let themeGlbSub = ["red", "skyblue", "black", "#00768F", "#AE1027", "#808C6C", "#0E97B6", "#EF5B9C", "#6C757D", "#D0D3DC"];
+let themeGlbCd = ["red", "skyblue", "black", "#FAC457", "#A12F2F", "#F4B46A", "#0E97B6", "#EF5B9C", "#6C757D", "#C8B493"];
+let themeGlbTxt = ["black", "black", "black", "#00768F", "#752423", "#9A9C94", "black", "#F7ACBC", "#6C757D", "#383C4A"];
+let themeGlbName = ["[官方]经典", "[官方]经典2", "[官方]护眼配色", "[投稿]游乐场", "[投稿]我都觉得好", "[投稿]春天来啦~", "[投稿]奶思", "[投稿]樱花粉", "[投稿]看个毛线", "[投稿]海盐夏日"];
 let themeLocBg = new Array(), themeLocSub = new Array(), themeLocCd = new Array(), themeLocTxt = new Array(), themeLocName = new Array();
 
 if (localStorage.getItem("theme_id") === null) localStorage.setItem("theme_id", "0");
@@ -33,6 +33,7 @@ chooseTheme(themeChosen);
 function getThemeLoc() {
 	if (localStorage.getItem("theme") === null) localStorage.setItem("theme", "");
 	let themeLoc = localStorage.getItem("theme").split(";");
+	themeLocName = [], themeLocBg = [], themeLocSub = [], themeLocCd = [], themeLocTxt = [];
 	for (let i = 0; i < themeLoc.length - 1; i++) {
 		themeLocName.push(themeLoc[i].split(",")[0]);
 		themeLocBg.push(themeLoc[i].split(",")[1]);
@@ -45,6 +46,8 @@ function addThemeLoc(name, bg, sub, cd, txt) {
 	localStorage.setItem("theme", localStorage.getItem("theme") + name + "," + bg + "," + sub + "," + cd + "," + txt + ";");
 }
 function delThemeLoc(n) {
+	let conf = confirm("确认删除该主题？");
+	if (!conf) return;
 	let themeLoc = localStorage.getItem("theme").split(";");
 	themeLoc.splice(n, 1);
 	localStorage.setItem("theme", themeLoc.join(";"));
@@ -69,7 +72,7 @@ for (let i = 0; i < themeGlbBg.length; i++) {
 for (let i = themeGlbBg.length; i < themeGlbBg.length + themeLocBg.length; i++) {
 	let thml = themeModelLoc;
 	thml = thml.replace(/ID/g, i.toString());
-	thml = thml.replace(/THEMENAME/g, themeLocName[i - themeGlbBg.length]);
+	thml = thml.replace(/THEMENAME/g, "[本地]" + themeLocName[i - themeGlbBg.length]);
 	document.getElementById("cd-tab5").innerHTML += thml;
 }
 for (let i = 0; i < 1000; i++) {
